refactor: 外移 Core 内联测试模块
